split user actions + updated imports
