+++ app/actions/user/get-wins.ts
@@ -1,0 +1,1 @@
+"use server"
